Rename components directory to pages, fix css typos
