+++ src/pages/Scan.vue
