+++ src/sass/app.scss
@@ -107,7 +107,7 @@ label {
 button {
   background-color: $arc_rose;
   color: $arc_white;
-  font-size:1em;
+  font-size: 1em;
   display: block;
   text-align: center;
   width: 70%;
@@ -156,7 +156,7 @@ header {
   margin: 0 auto;
   padding: 2em;
   max-width: 900px;
-  overflow:auto;
+  overflow: auto;
 }
 
 .content {
@@ -241,8 +241,6 @@ header {
       position: absolute;
       top: 0;
       left: 0;
-      -webkit-border-radius: 50%;
-      -moz-border-radius: 50%;
       border-radius: 50%;
     }
     &[type=radio] + label::after {
@@ -253,16 +251,12 @@ header {
       position: absolute;
       top: 9px;
       left: 9px;
-      -webkit-border-radius: 50%;
-      -moz-border-radius: 50%;
       border-radius: 50%;
       zoom: 1;
       filter: alpha(opacity=0);
       opacity: 0;
     }
     &[type=radio]:focus + label::before {
-      -webkit-box-shadow: 0 0 0 4px $arc_focus;
-      -moz-box-shadow: 0 0 0 4px $arc_focus;
       box-shadow: 0 0 0 4px $arc_focus;
     }
     &:checked + label::after {
